refactor: rename TreeLoader to ResourceHierarchyLoader
